@@ -1,9 +1,10 @@
+using Authorization.Core.Models;
 using Microsoft.AspNetCore.Identity.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore;
 
-namespace Authorization.Core
+namespace Authorization.Core.Infrastructure
 {
-    public partial class AuthenticationDbContext(DbContextOptions options) : IdentityDbContext(options)
+    public partial class AuthenticationDbContext(DbContextOptions options) : IdentityDbContext<User>(options)
     {
     }
 }
