@@ -1,16 +1,18 @@
-﻿namespace Authorization.SSO.Filters
+﻿using System.Collections.Generic;
+using System.Linq;
+using System.Net;
+using Authorization.Lib.Helpers;
+using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.Mvc.Filters;
+
+namespace Authorization.SSO.Filters
 {
     /// <summary>
     /// Request to this controller may be originated only from allowed IP addresses (ServerOptions.AllowedIPAddresses) 
     /// </summary>
-    internal class RequestAddressFilter : ActionFilterAttribute
+    internal class RequestAddressFilter(ServerOptions options) : ActionFilterAttribute
     {
-        private readonly List<string> ipList;
-
-        public RequestAddressFilter(ServerOptions options)
-        {
-            ipList = (options?.AllowedIPAddresses?.Count() ?? 0) > 0 ? options.AllowedIPAddresses.ToList() : null;
-        }
+        private readonly List<string> ipList = (options?.AllowedIPAddresses?.Count() ?? 0) > 0 ? options.AllowedIPAddresses.ToList() : null;
 
         public override void OnActionExecuting(ActionExecutingContext context)
         {
@@ -25,7 +27,7 @@
         private bool IsAllowed(ActionExecutingContext context)
         {
             var islocal = ipList?.Contains(context.HttpContext.Connection.RemoteIpAddress.ToString()) ?? true;
-            var claimId = context.HttpContext.User.Claims.FirstOrDefault(c => c.Type == AuthParams.AuthParams.idClaim);
+            var claimId = context.HttpContext.User.Claims.FirstOrDefault(c => c.Type == FgrTermsHelper.idClaim);
             var isServer = claimId.Value == "server";
             return islocal && isServer;
         }
