@@ -1,6 +1,6 @@
 ﻿// <auto-generated />
 using System;
-using Authorization.Core;
+using Authorization.Core.Infrastructure;
 using Authorization.SSO;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Infrastructure;
